refactor: :sparkles: addes typescript, svgr and mdi icons config
