+++ app/package.json
@@ -12,12 +12,17 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
-    "styled-components": "^5.3.5"
+    "styled-components": "^5.3.5",
+    "typescript": "^4.9.4",
+    "vite-plugin-svgr": "^2.4.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
+    "@iconify-json/mdi": "^1.1.44",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@types/styled-components": "^5.1.26",
     "@vitejs/plugin-react": "^2.0.0",
+    "unplugin-icons": "^0.15.2",
     "vite": "^3.0.0"
   }
 }
