+++ app/src/components/botoes/LinkCadastro/index.ts
